+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - `split-records` filter now allows specifying the meta-data field in which to store the split name
+- centralized comparison code in `ldc.api._comparison` (available via `ldc.api`)
 
 
 0.2.8 (2025-07-15)